Branch create functionality added

+++ src/main/java/com/ashish/cactus/school/admin/db/repositories/AdminUserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Component;
 import com.ashish.cactus.school.admin.db.model.User;
 
 @Component
-public interface AdminUserRepository extends CrudRepository <User, Long>{
+public interface AdminUserRepository extends CrudRepository <User, Integer>{
 
 	User findByUserName(String username);
 }
