+++ src/main/java/com/ashish/cactus/school/admin/db/repositories/SchoolMasterRepository.java
@@ -6,9 +6,8 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Component;
 
 import com.ashish.cactus.school.admin.db.model.SchoolMaster;
-import com.ashish.cactus.school.admin.db.model.User;
 
 @Component
-public interface SchoolMasterRepository extends CrudRepository <SchoolMaster, Long>{
+public interface SchoolMasterRepository extends CrudRepository <SchoolMaster, Integer>{
 	public List<SchoolMaster> findBySchoolCode(String schoolCode);
 }
